--- starter-code/src/components/nav/nav.js
@@ -1,15 +1,0 @@
-import React from "react";
-
-class Nav extends React.Component {
-  render() {
-      return (
-
-  
-        
-      );
-    }
-  }
-
-
-
-export default Nav;
